+++ nagios/autoload.php
@@ -5,7 +5,7 @@
  *
  * @author terbach <terbach@netbixx.com>
  * @license See licence file LICENCE.md
- * @version 1.0.0
+ * @version 1.1.0
  * @since 1.1.2
  * @copyright 2012 netbixx GmbH (http://www.netbixx.com)
  */
@@ -18,7 +18,10 @@ spl_autoload_register('_autoloader');
  * @param string $class
  */
 function _autoloader($class) {
-    set_include_path(get_include_path() . PATH_SEPARATOR . dirname(__FILE__) . '/lib/');
+    set_include_path(get_include_path() . PATH_SEPARATOR
+                     . dirname(__FILE__) . '/../lib/' . PATH_SEPARATOR
+                     . dirname(__FILE__) . '/../lib/Logging' . PATH_SEPARATOR
+                     . dirname(__FILE__) . '/../lib/Logging/Adapter');
 
     $tmp = explode ('\\', $class);
     $clazz = end($tmp);