+++ tests/index.php
@@ -4,21 +4,11 @@
  * The boostrap file which is run before the tests for loading the autoloader-function.
  *
  * @author terbach <terbach@netbixx.com>
- * @version 1.0.0
+ * @version 1.1.0
  * @since 1.0.0
  * @copyright 2012 netbixx GmbH (http://www.netbixx.com)
  *
  * @category tests
  */
-spl_autoload_register('_autoloader');
 
-function _autoloader($class) {
-    set_include_path(get_include_path() . PATH_SEPARATOR . dirname(__FILE__) . '/../lib/' . PATH_SEPARATOR . dirname(__FILE__) . '/source/');
-
-    $tmp = explode ('\\', $class);
-    $clazz = end($tmp);
-    $clazz = str_replace('_', '/', $clazz);
-    $requiredFile = (string) $clazz . '.php';
-
-    require_once $requiredFile;
-}
+require_once '../nagios/autoload.php';