+++ tests/source/PluginMockNoStatus.php
@@ -80,8 +80,8 @@ class PluginMockNoStatus extends Plugin {
         return parent::calcStatus($value);
     }
 
-    public function setStatusFlags($ok, $warn, $crit) {
-        parent::setStatusFlags($ok, $warn, $crit);
+    public function setStatusFlags($okValue, $warn, $crit) {
+        parent::setStatusFlags($okValue, $warn, $crit);
     }
 
     public function isOk() {